+++ src/term_image/__init__.py
@@ -20,12 +20,12 @@ Copyright (c) 2022
 
 from __future__ import annotations
 
-__all__ = ("set_font_ratio", "get_font_ratio", "FontRatio")
+__all__ = ("set_cell_ratio", "get_cell_ratio", "AutoCellRatio")
 __author__ = "AnonymouX47"
 
 from enum import Enum, auto
 from operator import truediv
-from typing import Union
+from typing import Optional, Union
 
 from .exceptions import TermImageError
 from .utils import get_cell_size
@@ -34,27 +34,26 @@ version_info = (0, 5, 0, "dev0")
 __version__ = ".".join(map(str, version_info))
 
 
-def get_font_ratio() -> float:
-    """Returns the global :term:`font ratio`.
+def get_cell_ratio() -> float:
+    """Returns the global :term:`cell ratio`.
 
-    See :py:func:`set_font_ratio`.
+    See :py:func:`set_cell_ratio`.
     """
     # `(1, 2)` is a fallback in case the terminal doesn't respond in time
-    return _font_ratio or truediv(*(get_cell_size() or (1, 2)))
+    return _cell_ratio or truediv(*(get_cell_size() or (1, 2)))
 
 
-def set_font_ratio(ratio: Union[float, FontRatio]) -> None:
-    """Sets the global :term:`font ratio`.
+def set_cell_ratio(ratio: Union[float, AutoCellRatio]) -> None:
+    """Sets the global :term:`cell ratio`.
 
     Args:
         ratio: Can be one of the following values.
 
-          * A positive ``float``: a fixed aspect ratio of a character cell in the
-            terminal emulator.
-          * :py:attr:`FontRatio.AUTO`: the ratio is immediately determined from the
-            :term:`active terminal`.
-          * :py:attr:`FontRatio.FULL_AUTO`: the ratio is determined from the
-            :term:`active terminal` whenever :py:func:`get_font_ratio` is called,
+          * A positive ``float`` value.
+          * :py:attr:`AutoCellRatio.FIXED`, the ratio is immediately determined from
+            the :term:`active terminal`.
+          * :py:attr:`AutoCellRatio.DYNAMIC`, the ratio is determined from the
+            :term:`active terminal` whenever :py:func:`get_cell_ratio` is called,
             though with some caching involved, such that the ratio is re-determined
             only if the terminal size changes.
 
@@ -62,7 +61,7 @@ def set_font_ratio(ratio: Union[float, FontRatio]) -> None:
         TypeError: An argument is of an inappropriate type.
         ValueError: An argument is of an appropriate type but has an
           unexpected/invalid value.
-        term_image.exceptions.TermImageError: Auto font ratio is not supported
+        term_image.exceptions.TermImageError: Auto cell ratio is not supported
           in the :term:`active terminal` or on the current platform.
 
     This value is taken into consideration when setting image sizes for **text-based**
@@ -70,45 +69,48 @@ def set_font_ratio(ratio: Union[float, FontRatio]) -> None:
     terminal.
 
     NOTE:
-        Changing the font ratio does not automatically affect any image that already
-        has it's size set. For a change in font ratio to have any effect, the image's
-        size has to be set again.
+        Changing the cell ratio does not automatically affect any image that has a
+        :term:`fixed size`. For a change in cell ratio to take effect, the image's
+        size has to be re-set.
 
     ATTENTION:
-        See :ref:`auto-font-ratio` for details about the auto modes.
+        See :ref:`auto-cell-ratio` for details about the auto modes.
     """
-    global _auto_font_ratio, _font_ratio
+    global _cell_ratio
 
-    if isinstance(ratio, FontRatio):
-        if _auto_font_ratio is None:
-            _auto_font_ratio = get_cell_size() is not None
+    if isinstance(ratio, AutoCellRatio):
+        if AutoCellRatio.is_supported is None:
+            AutoCellRatio.is_supported = get_cell_size() is not None
 
-        if not _auto_font_ratio:
+        if not AutoCellRatio.is_supported:
             raise TermImageError(
-                "Auto font ratio is not supported in the active terminal or on the "
+                "Auto cell ratio is not supported in the active terminal or on the "
                 "current platform"
             )
-        elif ratio is FontRatio.AUTO:
+        elif ratio is AutoCellRatio.FIXED:
             # `(1, 2)` is a fallback in case the terminal doesn't respond in time
-            _font_ratio = truediv(*(get_cell_size() or (1, 2)))
+            _cell_ratio = truediv(*(get_cell_size() or (1, 2)))
         else:
-            _font_ratio = None
+            _cell_ratio = None
     elif isinstance(ratio, float):
         if ratio <= 0.0:
             raise ValueError(f"'ratio' must be greater than zero (got: {ratio})")
-        _font_ratio = ratio
+        _cell_ratio = ratio
     else:
         raise TypeError(
-            f"'ratio' must be a float or FontRatio enum (got: {type(ratio).__name__})"
+            "'ratio' must be a float or AutoCellRatio enum member "
+            f"(got: {type(ratio).__name__})"
         )
 
 
-class FontRatio(Enum):
-    """Constants for auto font ratio modes"""
+class AutoCellRatio(Enum):
+    """Values for setting :ref:`auto-cell-ratio`."""
 
-    AUTO = auto()
-    FULL_AUTO = auto()
+    is_supported: Optional[bool]
+
+    FIXED = auto()
+    DYNAMIC = auto()
 
 
-_font_ratio = 0.5
-_auto_font_ratio = None
+_cell_ratio = 0.5
+AutoCellRatio.is_supported = None